Added requester interfaces. Moved endpoint interfaces to RiotSharp.Interfaces. Moved requester classes to RiotSharp.Http
