+++ RiotSharp/Http/Interfaces/IRateLimitedRequester.cs
@@ -1,0 +1,96 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RiotSharp.Http.Interfaces
+{
+    interface IRateLimitedRequester
+    {
+        /// <summary>
+        /// The number of allowed calls per 10 seconds
+        /// </summary>
+        int RateLimitPer10S { get; set; }
+        /// <summary>
+        /// The number of allowed calls per 10 minutes
+        /// </summary>
+        int RateLimitPer10M { get; set; }
+
+        /// <summary>
+        /// Create a get request and send synchronously it to the server.
+        /// </summary>
+        /// <param name="relativeUrl"></param>
+        /// <param name="region"></param>
+        /// <param name="addedArguments"></param>
+        /// <param name="useHttps"></param>
+        /// <returns>The content of the response.</returns>
+        /// <exception cref="RiotSharpException">Thrown if an Http error occurs. Contains the Http error code and error message.</exception>
+        string CreateGetRequest(string relativeUrl, Region region, List<string> addedArguments = null,
+            bool useHttps = true);
+
+        /// <summary>
+        ///  Create a get request and send asynchronously it to the server.
+        /// </summary>
+        /// <param name="relativeUrl"></param>
+        /// <param name="region"></param>
+        /// <param name="addedArguments"></param>
+        /// <param name="useHttps"></param>
+        /// <returns>The content of the response.</returns>
+        /// <exception cref="RiotSharpException">Thrown if an Http error occurs. Contains the Http error code and error message.</exception>
+        Task<string> CreateGetRequestAsync(string relativeUrl, Region region,
+            List<string> addedArguments = null, bool useHttps = true);
+
+        /// <summary>
+        /// Create a post request and send synchronously it to the server.
+        /// </summary>
+        /// <param name="relativeUrl"></param>
+        /// <param name="region"></param>
+        /// <param name="body"></param>
+        /// <param name="addedArguments"></param>
+        /// <param name="useHttps"></param>
+        /// <returns>The content of the response.</returns>
+        /// <exception cref="RiotSharpException">Thrown if an Http error occurs. Contains the Http error code and error message.</exception>
+        string CreatePostRequest(string relativeUrl, Region region, string body,
+            List<string> addedArguments = null, bool useHttps = true);
+
+        /// <summary>
+        /// Create a post request and send asynchronously it to the server.
+        /// </summary>
+        /// <param name="relativeUrl"></param>
+        /// <param name="region"></param>
+        /// <param name="body"></param>
+        /// <param name="addedArguments"></param>
+        /// <param name="useHttps"></param>
+        /// <returns>The content of the response.</returns>
+        /// <exception cref="RiotSharpException">Thrown if an Http error occurs. Contains the Http error code and error message.</exception>
+        Task<string> CreatePostRequestAsync(string relativeUrl, Region region, string body,
+            List<string> addedArguments = null, bool useHttps = true);
+
+        /// <summary>
+        /// Create a put request and send synchronously it to the server.
+        /// </summary>
+        /// <param name="relativeUrl"></param>
+        /// <param name="region"></param>
+        /// <param name="body"></param>
+        /// <param name="addedArguments"></param>
+        /// <param name="useHttps"></param>
+        /// <returns>The content of the response.</returns>
+        /// <exception cref="RiotSharpException">Thrown if an Http error occurs. Contains the Http error code and error message.</exception>
+        bool CreatePutRequest(string relativeUrl, Region region, string body, List<string> addedArguments = null,
+            bool useHttps = true);
+
+        /// <summary>
+        /// Create a post request and send asynchronously it to the server.
+        /// </summary>
+        /// <param name="relativeUrl"></param>
+        /// <param name="region"></param>
+        /// <param name="body"></param>
+        /// <param name="addedArguments"></param>
+        /// <param name="useHttps"></param>
+        /// <returns>The content of the response.</returns>
+        /// <exception cref="RiotSharpException">Thrown if an Http error occurs. Contains the Http error code and error message.</exception>
+        Task<bool> CreatePutRequestAsync(string relativeUrl, Region region, string body,
+            List<string> addedArguments = null, bool useHttps = true);
+    }
+}
