+++ RiotSharp/Http/Interfaces/IRequester.cs
@@ -1,0 +1,41 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net.Http;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RiotSharp.Http.Interfaces
+{
+    interface IRequester
+    {
+        /// <summary>
+        /// The api key. Required to for all requests.
+        /// </summary>
+        string ApiKey { get; set; }
+
+        /// <summary>
+        /// Create a get request and send it to the server.
+        /// </summary>
+        /// <param name="relativeUrl"></param>
+        /// <param name="rootDomain"></param>
+        /// <param name="addedArguments"></param>
+        /// <param name="useHttps"></param>
+        /// <returns>The content of the response.</returns>
+        /// <exception cref="RiotSharpException">Thrown if an Http error occurs. Contains the Http error code and error message.</exception>
+        string CreateGetRequest(string relativeUrl, string rootDomain, List<string> addedArguments = null,
+            bool useHttps = true);
+
+        /// <summary>
+        /// Create a get request and send it to the server asynchronously.
+        /// </summary>
+        /// <param name="relativeUrl"></param>
+        /// <param name="rootDomain"></param>
+        /// <param name="addedArguments"></param>
+        /// <param name="useHttps"></param>
+        /// <returns>The content of the response.</returns>
+        /// <exception cref="RiotSharpException">Thrown if an Http error occurs. Contains the Http error code and error message.</exception>
+        Task<string> CreateGetRequestAsync(string relativeUrl, string rootDomain,
+            List<string> addedArguments = null, bool useHttps = true);
+    }
+}
