+++ RiotSharp/SummonerEndpoint/SummonerBase.cs
@@ -1,5 +1,7 @@
 ﻿using Newtonsoft.Json;
 using RiotSharp.GameEndpoint;
+using RiotSharp.Http;
+using RiotSharp.Http.Interfaces;
 using RiotSharp.LeagueEndpoint;
 using RiotSharp.StatsEndpoint;
 using System.Collections.Generic;
@@ -31,7 +33,7 @@ namespace RiotSharp.SummonerEndpoint
 
         private const string IdUrl = "/{0}";
 
-        private RateLimitedRequester requester;
+        private IRateLimitedRequester requester;
         public Region Region { get; set; }
 
         internal SummonerBase()
@@ -40,7 +42,7 @@ namespace RiotSharp.SummonerEndpoint
         }
 
         //summoner base not default constructor
-        internal SummonerBase(string id, string name, RateLimitedRequester requester, Region region)
+        internal SummonerBase(string id, string name, IRateLimitedRequester requester, Region region)
         {
             this.requester = requester;
             Region = region;
