+++ RiotSharpTest/RateLimiterTest.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RiotSharp;
+using RiotSharp.Http;
 
 namespace RiotSharpTest
 {
